Feat: Initial UX and Dashboard Functions
In this commit, we created the dashboard page. Initially, this page displays a table with all the articles. To enhance user experience, I added filters by categories and a search bar. Additionally, I implemented an Edit button that opens a modal for editing articles. Furthermore, I restructured the global state to provide users with instant updates.